Added Skills Controller

+++ ResourceApplicationTool/Controllers/RolesController.cs
@@ -9,6 +9,7 @@ using System.Web.Mvc;
 using ResourceApplicationTool.Models;
 using ResourceApplicationTool.Utils;
 
+
 namespace ResourceApplicationTool.Controllers
 {
     public class RolesController : Controller
@@ -18,29 +19,23 @@ namespace ResourceApplicationTool.Controllers
         // GET: Roles
         public ActionResult Index()
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             ViewBag.userAccess = Session[Const.CLAIM.USER_ACCESS_LEVEL];
             return View(db.Roles.ToList());
         }
 
-        // GET: Roles/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Role role = db.Roles.Find(id);
-            if (role == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.userAccess = Session[Const.CLAIM.USER_ACCESS_LEVEL];
-            return View(role);
-        }
+       
 
         // GET: Roles/Create
         public ActionResult Create()
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             ViewBag.userAccess = Session[Const.CLAIM.USER_ACCESS_LEVEL];
             return View();
         }
@@ -52,6 +47,10 @@ namespace ResourceApplicationTool.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "RoleID,Name,JobDescription,AverageSalary")] Role role)
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Roles.Add(role);
@@ -66,6 +65,10 @@ namespace ResourceApplicationTool.Controllers
         // GET: Roles/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -87,6 +90,10 @@ namespace ResourceApplicationTool.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "RoleID,Name,JobDescription,AverageSalary")] Role role)
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(role).State = EntityState.Modified;
@@ -101,6 +108,10 @@ namespace ResourceApplicationTool.Controllers
         // GET: Roles/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -120,6 +131,10 @@ namespace ResourceApplicationTool.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!Common.CheckIfAdministrator(Session, User))
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
             Role role = db.Roles.Find(id);
             db.Roles.Remove(role);
             db.SaveChanges();
